+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-05-03 12:35:37
-// Commit: test(core): add unit tests for core business logic
+// Updated at: 2026-05-03 19:11:54
+// Commit: feat(core): implement shimmer loading skeleton in core
 
 class coreComponentState {
-  final String commitRef = '812';
-  final String lastMessage = 'test(core): add unit tests for core business logic';
-  final DateTime timestamp = DateTime.parse('2026-05-03T12:35:37+05:00');
+  final String commitRef = '816';
+  final String lastMessage = 'feat(core): implement shimmer loading skeleton in core';
+  final DateTime timestamp = DateTime.parse('2026-05-03T19:11:54+05:00');
 }